Fixed a null reference issue
When updating the inoutpair due to new prompt, the current bot id array was used by reference.
This issue was first spotted when adding new bot while inputing 3 new prompt from a new chat.
When the component first initializing, a brand new array was created based on current bot id. Thus the first 2 prompt did not trigger the issue.

The fix was to use the copy of bot id array instead of using the bot id array reference.